Refactor data validation module into a modular package structure with backward compatibility
